feat: fincube-7 User management service setup

+++ backend/user-management-service/src/mfs_business/mfs_business.controller.ts
@@ -1,34 +1,55 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller, Get, Post, Body, Put, Param, Delete, NotFoundException } from '@nestjs/common';
 import { MfsBusinessService } from './mfs_business.service';
-import { CreateMfsBusinessDto } from './dto/create-mfs_business.dto';
-import { UpdateMfsBusinessDto } from './dto/update-mfs_business.dto';
+
+import { MfsBusinessEntity } from './entities/mfs_business.entity';
 
 @Controller('mfs-business')
 export class MfsBusinessController {
-  constructor(private readonly mfsBusinessService: MfsBusinessService) {}
+  constructor(private readonly mfsBusinessService: MfsBusinessService) { }
 
   @Post()
-  create(@Body() createMfsBusinessDto: CreateMfsBusinessDto) {
-    return this.mfsBusinessService.create(createMfsBusinessDto);
+  async create(@Body() mfs_business_entity: MfsBusinessEntity): Promise<MfsBusinessEntity> {
+    return this.mfsBusinessService.create(mfs_business_entity);
+  }
+
+  @Post()
+  async findByEmail(@Body() email: string): Promise<MfsBusinessEntity> {
+    const mfs_business = await this.mfsBusinessService.findByEmail(email);
+    if (!mfs_business) {
+      throw new NotFoundException('MFS business not found');
+    } else {
+      return mfs_business;
+    }
   }
 
   @Get()
-  findAll() {
+  async findAll(): Promise<MfsBusinessEntity[]> {
     return this.mfsBusinessService.findAll();
   }
 
+
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.mfsBusinessService.findOne(+id);
+  async findOne(@Param('id') id: string): Promise<MfsBusinessEntity> {
+    const mfs_business = await this.mfsBusinessService.findOne(+id);
+    if (!mfs_business) {
+      throw new NotFoundException(`MFS business doesn't exist`);
+    } else {
+      return mfs_business;
+    }
   }
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateMfsBusinessDto: UpdateMfsBusinessDto) {
-    return this.mfsBusinessService.update(+id, updateMfsBusinessDto);
+  @Put(':id')
+  async update(@Param('id') id: string, @Body() mfs_business_entity: MfsBusinessEntity): Promise<MfsBusinessEntity> {
+    return this.mfsBusinessService.update(+id, mfs_business_entity);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.mfsBusinessService.remove(+id);
+  async remove(@Param('id') id: string): Promise<any> {
+    const mfs_business = await this.mfsBusinessService.findOne(+id);
+    if (!mfs_business) {
+      throw new NotFoundException(`MFS business doesn't exist`);
+    } else {
+      return this.mfsBusinessService.remove(+id);
+    }
   }
 }
